+++ docs/en/reference/types.rst
@@ -701,9 +701,7 @@ Please also notice the mapping specific footnotes for additional information.
 |                   |               +--------------------------+         |                                                          |
 |                   |               | **Oracle**               |         |                                                          |
 |                   |               +--------------------------+---------+----------------------------------------------------------+
-|                   |               | **SQL Anywhere**         | < 12    | ``DATETIME``  [14]_ [15]_                                |
-|                   |               |                          +---------+----------------------------------------------------------+
-|                   |               |                          | >= 12   | ``TIMESTAMP WITH TIME ZONE``                             |
+|                   |               | **SQL Anywhere**         | "all"   | ``TIMESTAMP WITH TIME ZONE``                             |
 +-------------------+---------------+--------------------------+---------+----------------------------------------------------------+
 | **time**          | ``\DateTime`` | **MySQL**                | *all*   | ``TIME``                                                 |
 |                   |               +--------------------------+         |                                                          |
