+++ lib/Doctrine/DBAL/Platforms/Keywords/SQLAnywhereKeywords.php
@@ -30,6 +30,7 @@ class SQLAnywhereKeywords extends KeywordList
             'ALTER',
             'AND',
             'ANY',
+            'ARRAY',
             'AS',
             'ASC',
             'ATTACH',
@@ -70,6 +71,7 @@ class SQLAnywhereKeywords extends KeywordList
             'CURRENT_USER',
             'CURSOR',
             'DATE',
+            'DATETIMEOFFSET',
             'DBSPACE',
             'DEALLOCATE',
             'DEC',
@@ -116,7 +118,6 @@ class SQLAnywhereKeywords extends KeywordList
             'IF',
             'IN',
             'INDEX',
-            'INDEX_LPAREN',
             'INNER',
             'INOUT',
             'INSENSITIVE',
@@ -129,20 +130,22 @@ class SQLAnywhereKeywords extends KeywordList
             'INTEGRATED',
             'INTERSECT',
             'INTO',
-            'IQ',
             'IS',
             'ISOLATION',
             'JOIN',
+            'JSON',
             'KERBEROS',
             'KEY',
             'LATERAL',
             'LEFT',
             'LIKE',
+            'LIMIT',
             'LOCK',
             'LOGIN',
             'LONG',
             'MATCH',
             'MEMBERSHIP',
+            'MERGE',
             'MESSAGE',
             'MODE',
             'MODIFY',
@@ -160,6 +163,8 @@ class SQLAnywhereKeywords extends KeywordList
             'OFF',
             'ON',
             'OPEN',
+            'OPENSTRING',
+            'OPENXML',
             'OPTION',
             'OPTIONS',
             'OR',
@@ -196,6 +201,8 @@ class SQLAnywhereKeywords extends KeywordList
             'RIGHT',
             'ROLLBACK',
             'ROLLUP',
+            'ROW',
+            'ROWTYPE',
             'SAVE',
             'SAVEPOINT',
             'SCROLL',
@@ -207,6 +214,7 @@ class SQLAnywhereKeywords extends KeywordList
             'SHARE',
             'SMALLINT',
             'SOME',
+            'SPATIAL',
             'SQLCODE',
             'SQLSTATE',
             'START',
@@ -214,7 +222,6 @@ class SQLAnywhereKeywords extends KeywordList
             'SUBTRANS',
             'SUBTRANSACTION',
             'SYNCHRONIZE',
-            'SYNTAX_ERROR',
             'TABLE',
             'TEMPORARY',
             'THEN',
@@ -224,6 +231,7 @@ class SQLAnywhereKeywords extends KeywordList
             'TO',
             'TOP',
             'TRAN',
+            'TREAT',
             'TRIGGER',
             'TRUNCATE',
             'TSEQUAL',
@@ -232,6 +240,7 @@ class SQLAnywhereKeywords extends KeywordList
             'UNIQUE',
             'UNIQUEIDENTIFIER',
             'UNKNOWN',
+            'UNNEST',
             'UNSIGNED',
             'UPDATE',
             'UPDATING',
@@ -243,6 +252,7 @@ class SQLAnywhereKeywords extends KeywordList
             'VARBIT',
             'VARCHAR',
             'VARIABLE',
+            'VARRAY',
             'VARYING',
             'VIEW',
             'WAIT',
@@ -252,9 +262,6 @@ class SQLAnywhereKeywords extends KeywordList
             'WHILE',
             'WINDOW',
             'WITH',
-            'WITH_CUBE',
-            'WITH_LPAREN',
-            'WITH_ROLLUP',
             'WITHIN',
             'WORK',
             'WRITETEXT',
